phsar/requirements.txt
requests==2.32.3
SQLAlchemy==2.0.40
pydantic-settings==2.9.1
pandas==2.2.3
fastapi==0.115.12
uvicorn==0.34.2
asyncpg==0.30.0
alembic==1.15.2
bcrypt==4.3.0
passlib==1.7.4
pgvector==0.4.0
httpx==0.28.1
tenacity==9.1.2